Add Register Form and registerUser service

+++ components/forms/register/RegisterForm.jsx
@@ -1,14 +1,16 @@
 import styles from "../SignForms.module.css"
+import Link from "next/link"
 import { useRouter } from 'next/router'
 import { useState } from "react"
-import Link from "next/link"
-import loginUser from "../../../services/users/loginUser"
+import registerUser from "../../../services/users/registerUser"
 
-const LoginForm = () => {
+const RegisterForm = () => {
     const router = useRouter()
     const [formData, setFormData] = useState({
         username: "",
+        email: "",
         password: "",
+        confirmPassword: "",
         error: "",
         isLoading: false,
         bounce: false
@@ -20,7 +22,17 @@ const LoginForm = () => {
 
         setFormData({ ...formData, isLoading: true, bounce: false })
 
-        loginUser(formData)
+        if (formData.password !== formData.confirmPassword) {
+            setFormData({
+                ...formData,
+                isLoading: false,
+                bounce: true,
+                error: "Les mots de passe ne correspondent pas"
+            })
+            return null
+        }
+
+        registerUser(formData)
             .then((res) => {
                 const token = res.data.token
                 localStorage.setItem("token", token)
@@ -31,6 +43,7 @@ const LoginForm = () => {
                 setFormData({
                     ...formData,
                     password: "",
+                    confirmPassword: "",
                     error: e.response.data.message || e.message,
                     isLoading: false,
                     bounce: true
@@ -42,54 +55,69 @@ const LoginForm = () => {
         <div className={styles.form_wrapper}>
             <form onSubmit={handleSubmit} className={`${styles.form} container p-x`}>
                 <div className={styles.headings}>
-                    <h2>Connection</h2>
+                    <h2>Inscription</h2>
                 </div>
 
                 {formData.error && <div id={"error"} className={`${formData.bounce ? "bounce" : null } ${styles.error}`}>{formData.error}</div>}
 
                 <input
                     value={formData.username}
-                    onChange={(e) => setFormData({ ...formData, username: e.target.value })}
+                    onChange={(e) => setFormData({ ...formData, username: e.target.value})}
                     type="text"
-                    name="username"
-                    id="username"
+                    name={"username"}
+                    id={"username"}
                     required={true}
-                    placeholder={"Adresse e-mail ou nom d'utilisateur"}
+                    placeholder={"Nom d'utilisateur (visible par tous)"}
                     autoComplete={"username"}
                 />
 
+                <input
+                    value={formData.email}
+                    onChange={(e) => setFormData({ ...formData, email: e.target.value})}
+                    type="email"
+                    name={"email"}
+                    id={"email"}
+                    required={true}
+                    placeholder={"Adresse e-mail"}
+                    autoComplete={"username"}
+                />
 
                 <input
                     value={formData.password}
-                    onChange={(e) => setFormData({ ...formData, password: e.target.value })}
+                    onChange={(e) => setFormData({ ...formData, password: e.target.value})}
                     type="password"
-                    name="password"
-                    id="password"
+                    name={"password"}
+                    id={"password"}
                     required={true}
                     placeholder={"Mot de passe"}
                     autoComplete={"password"}
+                />
 
+                <input
+                    value={formData.confirmPassword}
+                    onChange={(e) => setFormData({ ...formData, confirmPassword: e.target.value})}
+                    type="password"
+                    name={"confirmPassword"}
+                    id={"confirmPassword"}
+                    required={true}
+                    placeholder={"Confirmer le mot de passe"}
+                    autoComplete={false}
                 />
 
                 <div className={styles.links}>
                     <button disabled={formData.isLoading} type="submit" className={`${styles.login_btn} btn`}>
-                        {formData.isLoading ? "Envoi ..." : "Se connecter"}
+                        {formData.isLoading ? "Envoi ..." : "Créer le compte"}
                     </button>
                     <p>
-                        Pas encore de compte ?
-                        <Link href={"/register"}>
-                            <a>Créez-en un !</a>
+                        Vous avez déjà un compte ?
+                        <Link href={"/login"}>
+                            <a>Connectez-vous !</a>
                         </Link>
                     </p>
-
                 </div>
-
             </form>
         </div>
-
     )
-
-
 }
 
-export default LoginForm
+export default RegisterForm
